added final report link in Croplist

+++ FarmVerse/farmVerseApplication/src/main/java/edu/infosys/farmVerseApplication/bean/FarmCropInputs.java
@@ -11,7 +11,7 @@ public class FarmCropInputs {
 	  private Double waterGallon;
 	  private Double fertilizer;
 	  private Double pesticides;
-	  private Integer tractorHour;
+	  private Double tractorHour;
 	  
 	  
 	  
@@ -22,7 +22,7 @@ public class FarmCropInputs {
 	}
 	  public FarmCropInputs(String cropId, String cropName, Double cropArea, String soil, String sownMonthYear,
 			String harvestMonthYear, Double yield, Double waterGallon, Double fertilizer, Double pesticides,
-			Integer tractorHour) {
+			Double tractorHour) {
 		super();
 		this.cropId = cropId;
 		this.cropName = cropName;
@@ -49,7 +49,7 @@ public class FarmCropInputs {
 			this.waterGallon = 0.0;
 			this.fertilizer = 0.0;
 			this.pesticides = 0.0;
-			this.tractorHour = 0;
+			this.tractorHour = 0.0;
 		  }
 	  public String getCropId() {
 		  return cropId;
@@ -111,10 +111,10 @@ public class FarmCropInputs {
 	  public void setPesticides(Double pesticides) {
 		  this.pesticides = pesticides;
 	  }
-	  public Integer getTractorHour() {
+	  public Double getTractorHour() {
 		  return tractorHour;
 	  }
-	  public void setTractorHour(Integer tractorHour) {
+	  public void setTractorHour(Double tractorHour) {
 		  this.tractorHour = tractorHour;
 	  }
